+++ ydb/core/health_check/health_check_utils.h
@@ -5,11 +5,29 @@
 #include <ydb/core/protos/blobstorage_distributed_config.pb.h>
 #include <ydb/core/sys_view/common/events.h>
 #include <ydb/core/tx/scheme_cache/scheme_cache.h>
+#include <ydb/library/actors/interconnect/interconnect.h>
 #include <ydb/public/api/grpc/ydb_monitoring_v1.grpc.pb.h>
+
+///
+#include <ydb/core/base/path.h>
+#include <ydb/core/base/statestorage.h>
+#include <ydb/core/mon/mon.h>
+#include <ydb/core/blobstorage/base/blobstorage_events.h>
+#include <ydb/core/mind/tenant_slot_broker.h>
+#include <ydb/core/tx/schemeshard/schemeshard.h>
+#include <ydb/core/util/proto_duration.h>
+#include <ydb/core/util/tuples.h>
+
+#include <ydb/core/protos/config.pb.h>
+#include <ydb/library/actors/interconnect/interconnect.h>
+#include <ydb/library/actors/wilson/wilson_span.h>
+#include <ydb/library/wilson_ids/wilson.h>
 
 namespace NKikimr::NHealthCheck {
 
 using namespace NSchemeCache;
+using namespace NNodeWhiteboard;
+using namespace NSchemeShard;
 
 using TGroupId = ui32;
 
@@ -41,22 +59,22 @@ static TString GetVDiskId(const NKikimrBlobStorage::TVDiskID& protoVDiskId) {
             << protoVDiskId.vdisk();
 }
 
-static TString GetVDiskId(const NKikimrBlobStorage::TBaseConfig::TVSlot& protoVSlot) {
-    return TStringBuilder()
-            << protoVSlot.groupid() << '-'
-            << protoVSlot.groupgeneration() << '-'
-            << protoVSlot.failrealmidx() << '-'
-            << protoVSlot.faildomainidx() << '-'
-            << protoVSlot.vdiskidx();
-}
+// static TString GetVDiskId(const NKikimrBlobStorage::TBaseConfig::TVSlot& protoVSlot) {
+//     return TStringBuilder()
+//             << protoVSlot.groupid() << '-'
+//             << protoVSlot.groupgeneration() << '-'
+//             << protoVSlot.failrealmidx() << '-'
+//             << protoVSlot.faildomainidx() << '-'
+//             << protoVSlot.vdiskidx();
+// }
 
 static TString GetVDiskId(const NKikimrBlobStorage::TNodeWardenServiceSet_TVDisk& protoVDiskId) {
     return GetVDiskId(protoVDiskId.vdiskid());
 }
 
-static TString GetVDiskId(const NKikimrWhiteboard::TVDiskStateInfo vDiskInfo) {
-    return GetVDiskId(vDiskInfo.vdiskid());
-}
+// static TString GetVDiskId(const NKikimrWhiteboard::TVDiskStateInfo vDiskInfo) {
+//     return GetVDiskId(vDiskInfo.vdiskid());
+// }
 
 static TString GetVDiskId(const NKikimrSysView::TVSlotInfo& vSlot) {
     return TStringBuilder()
@@ -75,13 +93,13 @@ static TString GetPDiskId(const NKikimrWhiteboard::TPDiskStateInfo pDiskInfo) {
     return TStringBuilder() << pDiskInfo.nodeid() << "-" << pDiskInfo.pdiskid();
 }
 
-static TString GetPDiskId(const NKikimrBlobStorage::TBaseConfig::TPDisk& pDisk) {
-    return TStringBuilder() << pDisk.nodeid() << "-" << pDisk.pdiskid();
-}
+// static TString GetPDiskId(const NKikimrBlobStorage::TBaseConfig::TPDisk& pDisk) {
+//     return TStringBuilder() << pDisk.nodeid() << "-" << pDisk.pdiskid();
+// }
 
-static TString GetPDiskId(const NKikimrBlobStorage::TBaseConfig::TVSlot& vSlot) {
-    return TStringBuilder() << vSlot.vslotid().nodeid() << "-" << vSlot.vslotid().pdiskid();
-}
+// static TString GetPDiskId(const NKikimrBlobStorage::TBaseConfig::TVSlot& vSlot) {
+//     return TStringBuilder() << vSlot.vslotid().nodeid() << "-" << vSlot.vslotid().pdiskid();
+// }
 
 static TString GetPDiskId(const NKikimrBlobStorage::TNodeWardenServiceSet_TPDisk& pDisk) {
     return TStringBuilder() << pDisk.nodeid() << "-" << pDisk.pdiskid();
@@ -93,6 +111,29 @@ static TString GetPDiskId(const NKikimrSysView::TVSlotKey& vSlotKey) {
 
 static TString GetPDiskId(const NKikimrSysView::TPDiskKey& pDiskKey) {
     return TStringBuilder() << pDiskKey.GetNodeId() << "-" << pDiskKey.GetPDiskId();
+}
+
+static bool IsSuccess(const std::unique_ptr<TEvSchemeShard::TEvDescribeSchemeResult>& ev) {
+    return ev->GetRecord().status() == NKikimrScheme::StatusSuccess;
+}
+
+static TString GetError(const std::unique_ptr<TEvSchemeShard::TEvDescribeSchemeResult>& ev) {
+    return NKikimrScheme::EStatus_Name(ev->GetRecord().status());
+}
+
+static bool IsSuccess(const std::unique_ptr<TEvStateStorage::TEvBoardInfo>& ev) {
+    return ev->Status == TEvStateStorage::TEvBoardInfo::EStatus::Ok;
+}
+
+static TString GetError(const std::unique_ptr<TEvStateStorage::TEvBoardInfo>& ev) {
+    switch (ev->Status) {
+        case TEvStateStorage::TEvBoardInfo::EStatus::Ok:
+            return "Ok";
+        case TEvStateStorage::TEvBoardInfo::EStatus::Unknown:
+            return "Unknown";
+        case TEvStateStorage::TEvBoardInfo::EStatus::NotAvailable:
+            return "NotAvailable";
+    }
 }
 
 static bool IsSuccess(const std::unique_ptr<TEvTxProxySchemeCache::TEvNavigateKeySetResult>& ev) {
@@ -137,5 +178,42 @@ static TString GetError(const std::unique_ptr<TEvTxProxySchemeCache::TEvNavigate
     return "no error";
 }
 
+// static Ydb::Monitoring::StatusFlag::Status GetFlagFromBSPDiskSpaceColor(NKikimrBlobStorage::TPDiskSpaceColor::E flag) {
+//     switch (flag) {
+//         case NKikimrBlobStorage::TPDiskSpaceColor::GREEN:
+//         case NKikimrBlobStorage::TPDiskSpaceColor::CYAN:
+//             return Ydb::Monitoring::StatusFlag::GREEN;
+//         case NKikimrBlobStorage::TPDiskSpaceColor::LIGHT_YELLOW:
+//         case NKikimrBlobStorage::TPDiskSpaceColor::YELLOW:
+//             return Ydb::Monitoring::StatusFlag::YELLOW;
+//         case NKikimrBlobStorage::TPDiskSpaceColor::LIGHT_ORANGE:
+//         case NKikimrBlobStorage::TPDiskSpaceColor::PRE_ORANGE:
+//         case NKikimrBlobStorage::TPDiskSpaceColor::ORANGE:
+//             return Ydb::Monitoring::StatusFlag::ORANGE;
+//         case NKikimrBlobStorage::TPDiskSpaceColor::RED:
+//             return Ydb::Monitoring::StatusFlag::RED;
+//         default:
+//             return Ydb::Monitoring::StatusFlag::UNSPECIFIED;
+//     }
+// }
+
+static Ydb::Monitoring::StatusFlag::Status GetFlagFromWhiteboardFlag(NKikimrWhiteboard::EFlag flag) {
+    switch (flag) {
+        case NKikimrWhiteboard::EFlag::Green:
+            return Ydb::Monitoring::StatusFlag::GREEN;
+        case NKikimrWhiteboard::EFlag::Yellow:
+            return Ydb::Monitoring::StatusFlag::YELLOW;
+        case NKikimrWhiteboard::EFlag::Orange:
+            return Ydb::Monitoring::StatusFlag::ORANGE;
+        case NKikimrWhiteboard::EFlag::Red:
+            return Ydb::Monitoring::StatusFlag::RED;
+        default:
+            return Ydb::Monitoring::StatusFlag::UNSPECIFIED;
+    }
+}
+
+// static TString GetNodeLocation(const TEvInterconnect::TNodeInfo& nodeInfo) {
+//     return TStringBuilder() << nodeInfo.NodeId << '/' << nodeInfo.Host << ':' << nodeInfo.Port;
+// }
 
 } // NKikimr::NHealthCheck
