+++ ydb/core/health_check/ya.make
@@ -1,7 +1,6 @@
 LIBRARY()
 
 SRCS(
-    builder.cpp
     health_check.cpp
     health_check.h
 )
